refactor(app): move public routes into a (site) route group
Splits the Navbar out of the root layout into a (site)-scoped layout so
the admin section can render without it.
