+++ src/app/(site)/clubs/[clubId]/_components/PhotoLightbox.tsx
