+++ src/app/(site)/clubs/[clubId]/_components/index.ts
@@ -1,0 +1,5 @@
+export * from "@/app/(site)/clubs/[clubId]/_components/BackLink";
+export * from "@/app/(site)/clubs/[clubId]/_components/ClubAbout";
+export * from "@/app/(site)/clubs/[clubId]/_components/ClubContacts";
+export * from "@/app/(site)/clubs/[clubId]/_components/ClubGallery";
+export * from "@/app/(site)/clubs/[clubId]/_components/ClubHeader";
