+++ src/app/(site)/register/club/_components/StepIndicator.tsx
