+++ src/app/(site)/register/page.tsx
@@ -1,5 +1,5 @@
 import { registerGuard } from "@/server/guard";
-import { RegisterForm } from "@/app/register/_components/RegisterForm";
+import { RegisterForm } from "@/app/(site)/register/_components/RegisterForm";
 
 export default async function Register() {
   await registerGuard();
