--- src/app/clubs/[clubId]/_components/index.ts
@@ -1,5 +1,0 @@
-export * from "@/app/clubs/[clubId]/_components/BackLink";
-export * from "@/app/clubs/[clubId]/_components/ClubAbout";
-export * from "@/app/clubs/[clubId]/_components/ClubContacts";
-export * from "@/app/clubs/[clubId]/_components/ClubGallery";
-export * from "@/app/clubs/[clubId]/_components/ClubHeader";
