--- src/app/clubs/_components/index.ts
@@ -1,6 +1,0 @@
-export * from "@/app/clubs/_components/CategoryFilterRow";
-export * from "@/app/clubs/_components/ClubCard";
-export * from "@/app/clubs/_components/ClubFilterModal";
-export * from "@/app/clubs/_components/ClubList";
-export * from "@/app/clubs/_components/ClubSearchBar";
-export * from "@/app/clubs/_components/ClubSortSelect";
