--- src/app/clubs/_lib/index.ts
@@ -1,2 +1,0 @@
-export * from "@/app/clubs/_lib/club-list-params";
-export * from "@/hooks/use-media-query";
