+++ src/app/layout.tsx
@@ -1,11 +1,8 @@
 import { type Metadata } from "next";
 import { IBM_Plex_Sans_Thai, Sarabun } from "next/font/google";
-import { headers } from "next/headers";
 import "@/styles/globals.css";
 
-import { auth } from "@/server/auth";
 import { TRPCReactProvider } from "@/trpc/react";
-import { Navbar } from "@/components/Navbar";
 
 export const metadata: Metadata = {
   title: {
@@ -38,17 +35,10 @@ const plexThai = IBM_Plex_Sans_Thai({
   variable: "--font-ibm-plex",
 });
 
-export default async function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
-  const session = await auth.api.getSession({ headers: await headers() });
-
+export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
   return (
     <html lang="th" className={`${sarabun.variable} ${plexThai.variable}`}>
       <body>
-        <Navbar
-          isLoggedIn={!!session}
-          userName={session?.user.name}
-          userEmail={session?.user.email}
-        />
         <TRPCReactProvider>{children}</TRPCReactProvider>
       </body>
     </html>
